* Organize TabView as a ui component
